+++ src/main/java/region/jidogam/domain/jwt/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package region.jidogam.domain.jwt;
+package region.jidogam.domain.jwt.dto;
 
 public record TokenResponse (
     String accessToken
